🎨 add routes folder
